--- src/routes/communityChat.js
@@ -1,14 +1,0 @@
-const { Router } = require('express');
-const { CommunityChatController } = require('../controllers/communityChat');
-
-const router = Router();
-
-const communityChatController = new CommunityChatController();
-
-router.get('/all', communityChatController.getAll);
-router.get('/:id', communityChatController.getById);
-router.post('/create', communityChatController.create);
-router.put('/update/:id', communityChatController.update);
-router.delete('/delete/:id', communityChatController.exclude);
-
-module.exports = router;
