+++ src/routes/middleware/index.ts
@@ -1,9 +1,12 @@
-// const { NextFunction, Request, Response } = require('express');
-const jwt = require('jsonwebtoken');
-const userModel = require('../../models/user');
+import { NextFunction, Request, Response } from 'express';
+import jwt from 'jsonwebtoken';
+import { UserRepository } from '../../repositories/user';
+import { IUserToken } from '../../types/interface';
 
-class Middleware {
-    async auth(req, res, next) {
+const userRepository = new UserRepository();
+
+export class Middleware {
+    async auth(req: Request, res: Response, next: NextFunction) {
         const token = req.headers.authorization;
         try {
             if (!token) throw new Error('Token não fornecido');
@@ -16,9 +19,9 @@ class Middleware {
 
             if (!tokenHash) throw new Error('Token inválido');
 
-            const decoded = jwt.verify(tokenHash, process.env.JWT_SECRET);
+            const decoded = jwt.verify(tokenHash, process.env.JWT_SECRET) as IUserToken;
 
-            const  user = await userModel.getById(decoded.user_id);
+            const  user = await userRepository.getById(decoded.user_id);
 
             if (!user) throw new Error('Faça login para acessar sua conta');
 
@@ -30,7 +33,3 @@ class Middleware {
         }
     }
 }
-
-module.exports = {
-    Middleware,
-};
