+++ src/routes/user.ts
@@ -1,5 +1,5 @@
-const { Router } = require('express');
-const { UserController } = require('../controllers/user');
+import { Router } from 'express';
+import { UserController } from '../controllers/user';
 
 const router = Router();
 
@@ -11,4 +11,4 @@ router.post('/create', userController.create);
 router.put('/update/:id', userController.update);
 router.delete('/delete/:id', userController.exclude);
 
-module.exports = router;
+export default router;
